+++ front/src/App.js
@@ -1,12 +1,14 @@
 import React, { Component } from 'react';
 import './App.css';
-import { Switch, Route ,Redirect} from 'react-router-dom';
-import Inscription from'./component/Inscription'
+import { Switch, Route } from 'react-router-dom';
+
 import Identification from './component/Identification'
-import Auth from './component/Auth'
+
 import styled from 'styled-components';
 import backgroundImage from './asset/chef.jpeg'
-import App2 from './component/App2';
+import Home from './component/Home';
+import SignIn from './component/SignIn';
+import SignUp from './component/SignUp';
 
 
 
@@ -28,16 +30,14 @@ class App extends Component {
     }
   
   render() {
-    console.log("tutu",this.state.username[0])
-    // if (this.state.redirect === true) 
-    //   return <Redirect to="/app2" />
+    
     return (
       <AppFront>
         <Switch>
           <Route exact path="/" component={Identification} />
-          <Route path="/inscription" render={() => <Inscription handleSingingSuccess={this.setUsername}/>} />
-          <Route path="/auth" render={() => <Auth handleSingingSuccess={this.setUsername}/>} />
-          <Route  path="/app2" render={() => <App2 usernameObject={this.state.username}/>} /> 
+          <Route path="/signin" render={() => <SignIn handleSingingSuccess={this.setUsername}/>} />
+          <Route path="/signup" render={() => <SignUp handleSingingSuccess={this.setUsername}/>} />
+          <Route  path="/home" render={() => <Home usernameObject={this.state.username}/>} /> 
         </Switch>
       </AppFront>
     );
